fitkit/requirements.txt
streamlit>=1.35.0
opencv-python-headless>=4.9.0
mediapipe>=0.10.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
google-generativeai>=0.5.0
altair>=5.0.0
pyarrow>=14.0.0
